refactor: change styles, render product
