+++ src/App.css
@@ -1,10 +1,5 @@
 .App {
   text-align: center;
-  background-color: #fefab1;
-  background-image: url(./images/background.jpg);
-  background-position: bottom;
-  background-repeat: no-repeat;
-  background-size: 100vw;
 }
 
 .App-logo {
