+++ src/components/Layout.tsx
@@ -1,6 +1,6 @@
-import Footer from './Footer';
-import Header from './Header';
-import Main from './Main';
+import Footer from './footer/Footer';
+import Header from './header/Header';
+import Main from './main/Main';
 
 function Layout() {
   return (
