+++ src/components/main/Main.module.css
@@ -12,8 +12,8 @@
 }
 .search__and__sort {
   display: flex;
-  align-items: center;
-  justify-content: space-between;
+  justify-content: end;
+  gap: 50px;
 }
 .search__and__sort select {
   margin: 0;
@@ -25,7 +25,7 @@
 }
 .search {
   margin: 0;
-  width: calc(80% - 50px);
+  width: calc(50% - 50px);
   height: 40px;
   border: 1px #d3d3d3 solid;
   background-image: url(../images/glass.png);
@@ -134,6 +134,7 @@ input[type=range]::-moz-range-thumb {
   display: flex;
   flex-wrap: wrap;
   justify-content: space-between;
+  row-gap: 30px;
 }
 .product {
   margin: 0;
